feat(source-scan): logging, refactor, image with same `cargo-near` embedded, remove `--no-docker` flag (#144)
Following output was generated by using the branch of current pr with
[build local `cargo-near`
Dockerfile](https://github.com/dj8yfo/cargo-near-image/tree/0.x.x-builer-local-cargo-near)
=> [image](https://hub.docker.com/r/dj8yfo/sourcescan/tags) on this
[repo](https://github.com/dj8yfo/sample_no_workspace/tree/set_metadata)


![Screenshot_20240412_230234](https://github.com/near/cargo-near/assets/26653921/9014a2de-33f8-4ddb-9d7d-d9a78ddb43df)

![Screenshot_20240412_230436](https://github.com/near/cargo-near/assets/26653921/a6f3023b-4858-460c-bbf0-4cdc94ea5bae)

![Screenshot_20240412_230749](https://github.com/near/cargo-near/assets/26653921/36f7f64f-ae50-4620-aa35-81f4db54eed0)
