+++ .github/workflows/test.yml
@@ -66,7 +66,7 @@ jobs:
         run: sudo apt-get update && sudo apt-get install --assume-yes libudev-dev
 
       - name: Run tests
-        run: cargo test --workspace --verbose
+        run: cargo test --workspace
 
   lint:
     runs-on: ubuntu-latest
